Refactored MaatCourtDataService into separate classes
